Fix: Refactor get_device_doc with moved filter
