+++ python-avd/pyavd/get_device_doc.py
@@ -2,6 +2,7 @@
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
 from .constants import JINJA2_DOCUMENTAITON_TEMPLATE
+from .j2filters.add_md_toc import add_md_toc as filter_add_md_toc
 from .templater import Templar
 
 
@@ -17,10 +18,6 @@ def get_device_doc(structured_config: dict, add_md_toc: bool = False) -> str:
     Returns:
         Device documentation in Markdown format.
     """
-    # pylint: disable=import-outside-toplevel
-    from .vendor.j2.filter.add_md_toc import add_md_toc as filter_add_md_toc
-
-    # pylint: enable=import-outside-toplevel
 
     templar = Templar()
     result: str = templar.render_template_from_file(JINJA2_DOCUMENTAITON_TEMPLATE, structured_config)
