+++ python-avd/pyavd/j2filters/add_md_toc.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2023-2024 Arista Networks, Inc.
 # Use of this source code is governed by the Apache License 2.0
 # that can be found in the LICENSE file.
-from __future__ import absolute_import, division, print_function
+from __future__ import annotations
 
 import re
 from unicodedata import normalize
